Reorder, move content around to have better structure.
