+++ Jetchat/app/src/main/java/com/example/compose/jetchat/video/VideoPlayer.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.compose.jetchat.conversation
+package com.example.compose.jetchat.video
 
 import android.graphics.Bitmap
 import android.graphics.RectF
@@ -68,7 +68,6 @@ import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.net.toUri
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LifecycleEventEffect
-import androidx.media3.common.C
 import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
 import androidx.media3.common.util.UnstableApi
@@ -79,6 +78,8 @@ import androidx.media3.ui.compose.state.rememberPlayPauseButtonState
 import androidx.media3.ui.compose.state.rememberPresentationState
 import androidx.media3.ui.compose.state.rememberProgressStateWithTickInterval
 import com.example.compose.jetchat.R
+import com.example.compose.jetchat.conversation.BlurRegionSpec
+import com.example.compose.jetchat.conversation.SurfaceViewBlurHelper
 import kotlin.time.Duration.Companion.milliseconds
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
